requirements.txt
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
distro==1.7.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fonttools==4.32.0
glob2==0.7
ipykernel==6.12.1
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.2.2
jupyter-core==4.9.2
kiwisolver==1.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.28
natsort==8.1.0
nest-asyncio==1.5.5
numpy==1.22.3
opencv-python==4.5.5.64
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.11.2
PyMuPDF==1.19.6
pyodbc==4.0.32
pyparsing==3.0.7
PyPDF2==1.27.1
PySimpleGUI==4.59.0
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
pyzmq==22.3.0
six==1.16.0
stack-data==0.2.0
tabula-py==2.3.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.2.0
wcwidth==0.2.5
